requirements.txt
python-dotenv
pyyaml
fastapi
uvicorn
aiokafka
redis
pydantic
sqlalchemy